dtvkit-amlogic: last playback position remains taking effect on next play [1/1]
PD#SWPL-145757

Problem:
When the previous playback ends up with EOS, the next playback on
the same recording will exit quickly. This is because some states
of previous playback are not cleaned completely on exit and affect
subsequent playback.

Solution:
Clean the said states on EOS of playback.

Verify:
Tested pass on ohm

Change-Id: Ie72249ccbf7ceddfafc8b2e78c22500de25b943f
Signed-off-by: Wentao.MA <wentao.ma@amlogic.com>