dtvkit-amlogic: Fails to return to live when pressing STOP/PAUSE quickly [1/1]
PD#SWPL-149528

Problem:
It fails to return to live from timeshift when pressing PAUSE
quickly following STOP. One reason is stopTimeshiftRecording
fails to stop PVR playback due to it is not in STARTED state.
In fact, it might be in STARTING state, in which it should also
stop PVR playback. If it fails to stop it, later live play will
not work.

Solution:
Also stop PVR playback if it is in STARTING state.

Verify:
Tested pass on ohm

Change-Id: Ic034df47d112051b1e4b474f7059b9a7ceb99131
Signed-off-by: Wentao.MA <wentao.ma@amlogic.com>