dtvkit-amlogic: Slow Progress Bar Update During Playback [1/1]
PD#SWPL-189323

Problem:
The progress bar updates slowly during playback. The progress
retrieved by the upper layer through Player.getStatus is not the
latest and has a slight delay.

Solution:
When Player.getStatus is called, calculate a compensation value
based on the current time, progress update time, and playback speed
to offset the delay in the playback progress.

Verify:
Tested pass on ohm

Signed-off-by: Wentao.MA <wentao.ma@amlogic.com>
Change-Id: I37f8baa7c27977b83a1f3dfaf543c2070d3da1a7